activex: fix initial volume value
_i_volume is used to set a custom initial volume over the html embed
attribute or the Visual Studio object property window.
A new value will be only accepted if it's different to _i_volume. So if
you want to set it to 50, it will currently not work, because activex
initializes libvlc with 100 (default).

This patch will set _i_volume to 100 to match the real volume.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>